Update README.md
The "Boston's Pop-Up ShopTracker" is a Python-based software designed for store owners and consumers in Boston to locate pop-up stores. Opening pop-up stores is a pretty common strategy for many branded shop owners to test their new merchandises’ sales. They are usually open for a short period of time, mostly during the summer and the location varies every year. It could be a small street shop or could be part of a shopping mall. One of the common problems for consumers is that it is difficult to keep track of these pop-up stores on when and where they will be open, as well as the merchandise that will be available in those stores. The software offers various functions, including summarizing goods availability in specific areas, store hours and time periods, providing detailed information about goods, their prices, and sorting them based on shop categories. Customers can utilize this tool to find shop details (location, goods, hours, etc) and owners can use the data to optimize shopping experiences, identify trends and meet customer demand. By using this software, both shop owners and customers can enhance their pop-up stores shopping experiences in Boston.